pyproject.toml

[tool.poetry]
name = "mahiru_habits_coursework"
version = "0.7.0"
authors = ["name <email>"]
readme = "README.md"
description = 'Проект для управления привычками с использованием Django, Celery и Telegram Bot.'

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
ipython = "^8.31.0"
django = "^5.1.4"
pillow = "^11.0.0"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
stripe = "^11.4.1"
django-celery-beat = "^2.7.0"
celery = { extras = ["redis"], version = "^5.4.0" }
django-cors-headers = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.0"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-tests = 'tests.test_utils:main'

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов 
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта`
)
'''

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
